changes to script and database

+++ script.sh
@@ -1,29 +1,38 @@
 #! /bin/bash
 
 ssh -i ~/.ssh/id_rsa ubuntu@18.202.167.4 << EOF
-
-	git clone https://github.com/LukeBenson/install-scripts.git
-	cd install-scripts/
-	./docker_install.sh
-	./docker-compose-install.sh
-	cd ~
-	rm -rf install-scripts/
 	
-	## Clone my project
-	git clone https://github.com/SherazShahid017/practical-project-qa
-	cd practical-project-qa/
-	git checkout dev
+	###################### Install docker/docker-compose ######################
+	if [ ! -f practical-project-qa ]; then
+		git clone https://github.com/LukeBenson/install-scripts.git
+		cd install-scripts/
+		./docker_install.sh
+		./docker-compose-install.sh
+		cd ~
+		rm -rf install-scripts/
+	fi
 
-	## Setting up mysql and running the file
-	sudo apt-get update
-	sudo apt install mysql-client-core-5.7
+
+	################# Clone my project #################################
+	if [ ! -f practical-project-qa ]; then
+    	    git clone https://github.com/SherazShahid017/practical-project-qa
+            cd practical-project-qa/
+            git checkout dev
+	fi	
+
+
+	####################### Setting up mysql and running the file######################
+	if ! which mysql > /dev/null; then
+		sudo apt-get update
+		sudo apt install mysql-client-core-5.7 -y
+	fi
 
 	mysql -h terraform-20201122151632162900000001.cr9zze1q6zyi.eu-west-1.rds.amazonaws.com -P 3306 -u admin -proot1234
 	use testdb;
 	source ~/practical-project-qa/database/Create.sql;
 	exit
 
-	## Running tests
+	########################### Running tests ################################
 	sudo docker-compose up -d
 		
 	sudo docker exec backend bash -c "pytest tests/ --cov application"
